Изменил сцену sample добавив кор геймплей и добавил необходимые ресурсы

+++ Assets/Scripts/WareHouse/Views/EnergyWareHouseView.cs
@@ -12,6 +12,6 @@ public class EnergyWareHouseView : AbstractWareHouseView
     }
     protected override void ChangeView()
     {
-        storageContent.value = (energyWareHouse.CurrentCountResources / energyWareHouse.MaxCountResources);
+        storageContent.value = ((float)energyWareHouse.CurrentCountResources / (float)energyWareHouse.MaxCountResources);
     }
 }
